+++ ASCOfficeRtfFile/RtfFormatLib/source/RtfGlobalTables.h
@@ -119,7 +119,7 @@ public:
 		
 		 if( true == RtfColor::GetThemeByString(sTheme,oTheme  ) )
 		 {
-			 for (size_t i = 0; i < (int)m_aArray.size(); i++ )
+			 for (size_t i = 0; i < m_aArray.size(); i++ )
 			 {
 				if( oTheme == m_aArray[i].m_eTheme )
 				{
@@ -132,11 +132,11 @@ public:
 	}
 	bool GetColor( RtfColor oColor , int & nId)
 	 {
-		 for (int i = 0; i < (int)m_aArray.size(); i++ )
+		 for (size_t i = 0; i < m_aArray.size(); i++ )
 		 {
 			 if( m_aArray[i] == oColor )
 			 {
-				nId = i + 1;
+				nId = (int)i + 1;
 				return true;
 			 }
 		 }
@@ -151,7 +151,7 @@ public:
 			RenderParameter oNewParameter = oRenderParameter;
 			oNewParameter.nType = RENDER_TO_RTF_PARAM_COLOR_TBL;
 			
-			for (size_t i = 0; i < (int)m_aArray.size(); i++ )
+			for (size_t i = 0; i < m_aArray.size(); i++ )
             {
                 sResult += m_aArray[i].RenderToRtf( oNewParameter );
             }
@@ -375,7 +375,7 @@ public:
 	 }
 	bool GetList( int nId, RtfListProperty& oListProperty )
 	 {
-		 for (size_t i = 0; i < (int)m_aArray.size(); i++ )
+		 for (size_t i = 0; i < m_aArray.size(); i++ )
 		 {
 			 if( nId == m_aArray[i].m_nID )
 			 {
@@ -417,7 +417,7 @@ class RtfListOverrideTable : public IDocumentElement, public ItemContainer<RtfLi
 public: 
 	bool GetList( int nId, RtfListOverrideProperty& oListOverrideProperty )
 	{
-		for (size_t i = 0; i < (int)m_aArray.size(); i++ )
+		for (size_t i = 0; i < m_aArray.size(); i++ )
 		{
 			 if( nId == m_aArray[i].m_nIndex )
 			 {
